requirements.txt
setuptools
torch
numpy
pydub
shazamio
spotipy
noisereduce
scipy
python-dotenv